calle nitpicks crud
